+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Health: Check if `vim.g.rustaceanvim` is set,
+  but hasn't been sourced before initialization.
+
 ## [3.7.1] - 2023-11-28
 
 ### Fixed